00554: If accounts is null, assume everything is null as well

+++ hedera-node/src/main/java/com/hedera/services/ServicesState.java
@@ -243,6 +243,10 @@ public class ServicesState extends AbstractMerkleInternal implements SwirldState
 	@Override
 	protected synchronized void onDelete() {
 		super.onDelete();
+		if (storage() == null) {
+			return;
+		}
+
 		storage().delete();
 		accounts().delete();
 		topics().delete();
